+++ docs/changelog.rst
@@ -2,6 +2,14 @@
  importlib_resources NEWS
 ==========================
 
+v1.5.0
+======
+
+* Traversable is now a Protocol instead of an Abstract Base
+  Class (Python 2.7 and Python 3.8+).
+
+* Traversable objects now require a ``.name`` property.
+
 v1.4.0
 ======
 * #79: Temporary files created will now reflect the filename of